+++ src/apis/my/my.type.ts
@@ -1,0 +1,11 @@
+//api
+export interface UserInfoPutDataProp {
+  newDepartment: string;
+  newUserName: string;
+  newUserNumber: string;
+}
+
+export interface UserPasswordPutProp {
+  password: string;
+  newPassword: string;
+}
